+++ ICInputAccessoryUITests/ICKeyboardDismissTextFieldUITests.swift
@@ -28,40 +28,24 @@ import XCTest
 
 class ICKeyboardDismissTextFieldUITests: XCTestCase {
 
+  private lazy var app = XCUIApplication()
+
   override func setUp() {
     super.setUp()
-    // In UI tests it is usually best to stop immediately when a failure occurs.
     continueAfterFailure = false
-    // UI tests must launch the application that they test. Doing this in setup will make sure it happens for each test method.
     XCUIApplication().launch()
   }
 
-  override func tearDown() {
-    // Put teardown code here. This method is called after the invocation of each test method in the class.
-    super.tearDown()
-  }
-
   func testKeyboardDismissing() {
-    let app = XCUIApplication()
-    app.tables.cells.textFields["ICKeyboardDismissTextField"].tap()
-
-    let keyboardWindow = app.children(matching: .window).element(boundBy: 1)
-    let accessory = keyboardWindow.children(matching: .other).element.children(matching: .other).element.children(matching: .other).element(boundBy: 0)
-    accessory.children(matching: .button).element.tap()
+    app.tables.textFields["ICKeyboardDismissTextField"].tap()
+    app.buttons["Dismiss Keyboard"].tap()
   }
 
   func testStoryboard() {
-    let app = XCUIApplication()
-    let tablesQuery = app.tables
-
-    tablesQuery.buttons["Storyboard"].tap()
-    tablesQuery.textFields["Storyboard ICKeyboardDismissTextField"].tap()
-
-    let keyboardWindow = app.children(matching: .window).element(boundBy: 1)
-    let accessory = keyboardWindow.children(matching: .other).element.children(matching: .other).element.children(matching: .other).element(boundBy: 0)
-    accessory.children(matching: .button).element.tap()
-
-    tablesQuery.buttons["Back to Code"].tap()
+    app.tables.buttons["Storyboard"].tap()
+    app.tables.textFields["Storyboard ICKeyboardDismissTextField"].tap()
+    app.buttons["Dismiss Keyboard"].tap()
+    app.tables.buttons["Back to Code"].tap()
   }
 
 }